requirements.txt
# GRDK Dependencies
# Includes all grdl dependencies (inherited), explicitly defined

# =====================
# Core — grdl inherited
# =====================

# Numerical computing (used by both grdl and grdk throughout)
numpy>=1.20.0

# Scientific computing — interpolation, ndimage filters
# Used by grdl: coregistration, geolocation, imagej modules, ortho
scipy>=1.7.0

# =====================
# Core — grdk direct
# =====================

# Sibling library (image processing, IO, geolocation, coregistration)
# Used by: core/executor.py, core/discovery.py, widgets/geodev/*, __main__.py
grdl>=0.1.0

# YAML workflow definitions (grdk/core/dsl.py)
PyYAML>=6.0

# Artifact update checks against PyPI/conda (grdk/catalog/updater.py)
requests>=2.25.0

# Version comparison for artifact updates (grdk/catalog/updater.py)
packaging>=21.0

# =====================
# GUI Framework
# =====================

# Qt bindings — all widget and viewer modules
# Used by: widgets/admin/*, widgets/geodev/*, widgets/_display_controls.py,
#          widgets/_param_controls.py, viewers/image_canvas.py, viewers/chip_gallery.py,
#          viewers/stack_viewer.py
PyQt6>=6.5

# Orange Canvas widget framework — all widget modules
# Used by: widgets/admin/*, widgets/geodev/*
orange-widget-base

# =====================
# Optional — grdl format readers (deferred imports from grdk widgets)
# =====================

# SICD/CPHD format support (widgets/geodev/ow_image_loader.py → grdl.IO.sar)
sarpy>=1.3.0

# GeoTIFF/Raster format support (widgets/geodev/ow_image_loader.py → grdl.IO.biomass, grdl.IO.sar)
rasterio>=1.2.0

# Feature-based co-registration (widgets/geodev/ow_coregister.py → grdl.coregistration)
opencv-python-headless>=4.5.0

# =====================
# Optional — GPU acceleration (conditional imports in grdk/core/gpu.py)
# Uncomment for CUDA-accelerated processing
# =====================

# cupy-cuda12x>=13.0
# torch>=2.0

# =====================
# Optional — Stack Viewer (deferred import in grdk/viewers/stack_viewer.py)
# =====================

# napari>=0.4
